tests: Move make_tm helper function to utils file
Preparation to be able to use the function from different test files later on.
