+++ test/utils.hpp
@@ -1,0 +1,8 @@
+/* Icinga 2 | (c) 2025 Icinga GmbH | GPLv2+ */
+
+#pragma once
+
+#include <ctime>
+#include <string>
+
+tm make_tm(std::string s);
